refactor: rename misleadingly-named view files and classes
Addresses code review feedback:
- applications/views/hacker.py → application.py (contains dispatch for hacker, volunteer, and mentor; not hacker-only)
- class HackerDashboard → ApplicationDashboard
- class HackerApplication → ApplicationEditView (also avoids shadowing the HackerApplication model)
- organizers/views/batch_ops.py → invite.py (file contains only InviteTeamListView)
- organizers/views/lists.py → application_lists.py (more specific than "lists")
