+++ applications/urls.py
@@ -8,8 +8,8 @@ urlpatterns = [
         name='confirm_app'),
     url(r'^applications/(?P<id>[\w-]+)/cancel$', login_required(views.CancelApplication.as_view()),
         name='cancel_app'),
-    url(r'^dashboard/$', views.HackerDashboard.as_view(), name='dashboard'),
-    url(r'^application/$', views.HackerApplication.as_view(), name='application'),
+    url(r'^dashboard/$', views.ApplicationDashboard.as_view(), name='dashboard'),
+    url(r'^application/$', views.ApplicationEditView.as_view(), name='application'),
     url(r'^application/change_to_mentor/$', views.ConvertHackerToMentor.as_view(), name='change_to_mentor'),
     url(r'^application/draft/$', views.save_draft, name='save_draft'),
     url(r'^sponsor/(?P<uid>[0-9A-Za-z_\-]+)/(?P<token>[\w-]+)/$',
