+++ applications/views/__init__.py
@@ -1,4 +1,4 @@
-from applications.views.hacker import (
+from applications.views.application import (
     VIEW_APPLICATION_TYPE,
     VIEW_APPLICATION_FORM_TYPE,
     check_application_exists,
@@ -6,8 +6,8 @@ from applications.views.hacker import (
     user_is_in_blacklist,
     ConfirmApplication,
     CancelApplication,
-    HackerDashboard,
-    HackerApplication,
+    ApplicationDashboard,
+    ApplicationEditView,
 )
 from applications.views.sponsor import SponsorApplicationView, SponsorDashboard
 from applications.views.mentor import ConvertHackerToMentor
@@ -16,7 +16,7 @@ from applications.views.draft import save_draft
 __all__ = [
     'VIEW_APPLICATION_TYPE', 'VIEW_APPLICATION_FORM_TYPE',
     'check_application_exists', 'get_deadline', 'user_is_in_blacklist',
-    'ConfirmApplication', 'CancelApplication', 'HackerDashboard', 'HackerApplication',
+    'ConfirmApplication', 'CancelApplication', 'ApplicationDashboard', 'ApplicationEditView',
     'SponsorApplicationView', 'SponsorDashboard',
     'ConvertHackerToMentor',
     'save_draft',
